distrib-nightly-android.sh now relies on two env variables, $FBCARM and $SDLANDROID rather than hard-coding paths on James's machine
git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8910 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ distrib-nightly-android.sh
@@ -1,15 +1,29 @@
 #!/bin/sh
 
-# Paths in this are specific to James's computer, but it works, and is something to start with
+if [ ! -f "${FBCARM}" ] ; then
+  echo "The FBCARM env variable should point to the fbc compiler for arm"
+  exit 1
+fi
 
-cd ~/src/ohr/wip
-scons fbc=$FBCARM debug=0 android-source=1 game
-scons
-cd ~/src/sdl-android/project/jni/application
+if [ ! -d "${SDLANDROID}" ] ; then
+  echo "The SDLANDROID env variable should point to the checked out copy of sdl-android ohrrpgce branch"
+  exit 1
+fi
+
+SCRIPTDIR="${0%/*}"
+
+cd "${SCRIPTDIR}"
+scons fbc="${FBCARM}" debug=0 android-source=1 game
+cd "${SDLANDROID}"/project/jni/application
+
+# Make sure we are on the ohrrpgce branch
 git checkout ohrrpgce
+
+# update the src link to point to the ohrrpgce project
 rm src
 ln -s ohrrpgce src
-cd ~/src/sdl-android
+
+cd "${SDLANDROID}"
 rm project/bin/MainActivity-debug.apk
 ./build.sh
